Explicitly passing the project's root dir to the build_bootstrap.php script
This takes advantage of a recent change in the SensioDistributionBundle which allows this. A specific situation this addresses is when the `vendor` dir is symlinked, and the build_bootstrap.php script mistakenly places the bootstrap.php.cache in ../app of the vendor dir's real location (instead of actually inside the project).