bin/vendors: speed up: fetch only when necessary
If a bundle's HEAD is already at the version specified in deps or
deps.lock, don't fetch more commits from upstream.

This performance improvement only helps when the version specified is an
SHA-1 checksum (as opposed to a branch or tag name), but that turns out
to be quite often (especially if update/deps.lock is used).

This does not handle the case where upstream contains modified history
(e.g. missing commits), but neither does the rest of bin/vendors.